requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy>=2.0.36
bcrypt==4.1.1
Werkzeug==3.0.1
Pillow>=10.3.0